Enhance the navigation bar and Add axios map file
