--- src/assets/js/libraries.js
@@ -1,1 +1,0 @@
-export const _ = selector => document.querySelector(selector)
