+++ src/componets/home/login-form.js
@@ -3,114 +3,93 @@ import Base from './../Base.js'
 export default class LoginForm extends Base {
 
     css = `
-    .form {
-        z-index: 100;
-        position: absolute;
-        display: flex;
-        flex-direction: column;
-        text-align: center;
-        top: 50%;
-        left: 50%;
-        transform : translate(-50%, -50%);
-        width: 25%;
-        margin: 1em auto;
-        background-color: rgba(0,0,0,0.9);
-        color: #eeeeee;
-        padding: 0.5em 2em;
-        padding-bottom: 3em;
-        border-radius: 1px;        
+
+    #backdrop {
+        position: fixed;
+        z-index: 99;
+        background-color: rgba(0,0,0,0.7);
+        right: 0;
+        left: 0;
+        top: 0;
+        bottom: 0;
+        cursor: url(./assets/icon/remove-icon.png), auto;
     }
 
-    .row {
-        display: inherit;
+    .form {
+        position: relative;
+        z-index: 100;
+        width: 30%;
+        margin: auto;
+        background-color: rgba(0,0,0,0.9);
+        color: #eeeeee;
+        padding: 0.5rem 3rem;
+        padding-bottom: 3rem;
+        border-radius: 1px;
+        transition: all 0.5s ease;        
     }
 
     h2 {
-        margin: 1em;
+        margin: 1rem;
+        text-align: center;
     }
 
-    #backdrop {
-        position: absolute;
-        left: 0;
-        z-index: 99;
-        background-color: rgba(0,0,0,0.7);
-        width: 100%;
-        height: 100%;
-        padding: 0px; 
-        margin: 0px; 
-        padding-bottom: 80px;
-        cursor: pointer;
-    }
-
-    input {
+    .textField {
         outline: none;
         width: 80%;
         background-color: transparent;
         border: none;
         border-bottom: solid 1.25px #cccccc;
-        margin-bottom: 2.5em;
         color: #ffffff;
+        display: block;
+        margin: 2rem auto;
     }
     
-    input:focus,
-    input:valid {
+    .textField:focus,
+    .textField:valid {
         border-color: #38ee17;
     }
     
-    input[type=checkbox] {
+    .textField[type=checkbox] {
         display: inline;
     }
 
-    label {
+    .textField-label {
         position: absolute;
+        display: block;
         pointer-events: none;
-        left: 5.5vw;
+        transform: translate(3vw, -3.5rem);
         transition: all 0.2s;
     }
 
-    input:focus + label,
-    input:valid + label {
-        transform: translateY(-1.5em);
+    .textField:focus + .textField-label,
+    .textField:valid + .textField-label {
+        transform: translate(2rem, -4.5rem);
         font-size: 0.8em;
     }
 
+    
     .remember {
-        display: inline-grid;
-        grid-template-columns: auto auto;
+        text-align: center;
     }
-
-    .remember > input {
-        margin-right: 2vw;
-    }
-
-    .remember > span {
-        margin-left: 1vw;
-    }
-
-    .hr-separator {
-        width: 100%;
-        height: 2px;
-        background-color: #eeeeee;
-        margin: 2em 0;
+    
+    .remember > label {
+        cursor: pointer;
     }
 
     .remember input {
-        position: absolute;
-        opacity: 0;
+        display: none;
         cursor: pointer;
     }
 
     .checkmark {
+        position: absolute;
         height: 20px;
         width: 20px;
-        background-color:transparent;
+        background-color: transparent;
         border-radius: 50%;
         border: solid 1px #ffffff;
         transition: all 1s;
-    }
-
-    .remember:hover input ~ .checkmark {
-        background-color: #ccffcc;
+        transform: translateX(-2rem);
     }
 
     .remember input:checked ~ .checkmark {
@@ -135,115 +114,56 @@ export default class LoginForm extends Base {
         transform: rotate(45deg);
     }
 
-    button {
-        width: 90%;
-        display: inline-block;
-        height: 3em;
-        border-radius: 25px;
-        outline: none;
-        border: none;
-        background-image: linear-gradient(to right, #32be8f, #38d39f, #32be8f);
-        font-size: 0.8rem;
-        color: #fff;
-        text-transform: uppercase;
-        margin: 1em 0;
-        cursor: pointer;
-        transition: all 1s;
-    }
-        
-    button:disabled {
-        background-image: linear-gradient(to right, gray, gray);
-        background-color: gray;
-    } 
-
-    button:hover{
-        background-position: right;
-        color: black;
-    }
-
-    .google {
-        color: blue;
-        background-size: 200%;
-        background-image: url("../assets/img/google.svg");
-    }
-    
-
-    .facebook {
-        color: red;
-        background-size: 200%;
-        background-image: url("../assets/img/facebook.svg");
-    }
-
-    a {
-        cursor: pointer;
-        width: 100%;
-        text-align: right;
-        text-decoration: none;
-        color: #999;
-        font-size: 0.9rem;
-        transition: .3s;
-    }
-
-    a:hover{
-        color: #F4D03F;
-    }
-
-    .validation {
-        display: block;
-        color: red;
-        font-size: 0.8em;
-        font-weight: bold;
-    }
-
     @media screen and (max-width: 1200px) {
-        .form {
-            width: 30%;
+        .textField-label {
+            transform: translate(4vw, -3.5rem);
         }
-      }
+    }
 
     @media screen and (max-width: 992px) {
         .form {
             width: 40%;
         }
-      }
+    }
 
     @media screen and (max-width: 768px) {
         .form {
-            width: 80%;
+            width: 50%;
         }
-      }
+    }
+
+    @media screen and (max-width: 512px) {
+        .form {
+            width: 60%;
+        }
+    }
 
 `
     content = `
-    <div id="backdrop" title="Click to close this form">
+    <div id="backdrop" title=">>Click to close this form">
     </div>
 
     <div class="form">
         <h2>Login</h2>
         <div class="container">
-            <div class="row">
-                <input type="text" id="email" name="email" title="Email : someone@somthing.com" required />
-                <label for="email" id="email-label">Email</label>
+            <input class="textField" type="text" id="email" name="email" title="Email : someone@somthing.com" required />
+            <label class="textField-label" for="email" id="email-label">Email</label>
+        
+            <input class="textField" type="password" id="password" name="password" title= "Password : pass@123" required />
+            <label class="textField-label" for="password" id="password-label">Password</label>
+        
+            <div class="remember">
+                <input type="checkbox" id="remember" />
+                <label for="remember" class="checkmark"></label>
+                <label for="remember">Remember me</label>
             </div>
-            <div class="row">
-                <input type="password" id="password" name="password" title= "Password : pass@123" required />
-                <label for="password" id="password-label">Password</label>
-            </div>
-            <div class="row">
-                <div class="remember">
-                    <input type="checkbox" id="remember">
-                    <span class="checkmark"></span>
-                    <span>Remember me</span>
-                </div>
-            </div>
-            <div class="row">
-                <span class="validation" id="validation-email"></span>
-                <span class="validation" id="validation-password"></span>
-            </div>
-            <div class="row">
+        
+            <span class="validation" id="validation-email"></span>
+            <span class="validation" id="validation-password"></span>
+            <div>
                 <button id="login" disabled> Login </button>
             </div>
-            <div class="row">
+            <div>
                 <a title="Reset Password" id="reset-password">Forgot Password ? </a>
                 |
                 <a title="Create new Account" id="signup"> Sign Up </a>
@@ -252,14 +172,14 @@ export default class LoginForm extends Base {
             <div class="hr-separator">
             </div>
 
-            <div class="row">
+            <div>
                 <span>or</span>
             </div>
             
-            <div class="row">
+            <div>
                 <button class="google"><img class="img2" src="../assets/img/google.svg">Continue with Google</button>
             </div>
-            <div class="row">
+            <div>
                 <button class="facebook"><img class="img2" src="../assets/img/facebook.svg">Continue with Facebook</button>
             </div>
 
@@ -270,113 +190,11 @@ export default class LoginForm extends Base {
     constructor() {
         super()
         this.mount()
-
-        if (!(this.state.exit == null || this.state.exit)) {
-            dispatchEvent(new Event('exit-form'))
-            this.state.exit = true
-        }
+        
     }//End of constructor
 
     connectedCallback() {
 
-        this._qs('#backdrop').addEventListener('click', () => {
-            dispatchEvent(new Event('exit-form'))
-            this.setPath('/')
-        })
-
-        this._qs('#signup').addEventListener('click', () => dispatchEvent(new Event('signup-form')))
-
-        this._qs('#reset-password').addEventListener('click', () => dispatchEvent(new Event('reset-password-form')))
-
-        // Client side form validation
-        const validation = () => {
-            this.state.validation = false
-            const events = ['focus', 'keyup']
-            events.forEach(element => {
-                this._qs("#email").addEventListener(element, () => {
-                    if (this._qs("#email").value == '') {
-                        this._qs('#validation-email').innerHTML = "❌ Enter the email or Mobile"
-                        this.state.validation = false
-                    } else if (!(/^\w{2,}@\w{2,}\.\w{2,4}$/.test(this._qs("#email").value) || /^(?:7|0|(?:\+94))[0-9]{9,10}$/.test(this._qs("#email").value))) {
-                        this._qs('#validation-email').innerHTML = "❌ Enter a valid email or Mobile"
-                        this.state.validation = false
-                    } else {
-                        this._qs('#validation-email').innerHTML = ""
-                        this.state.validation = true
-                    }
-
-                    this.state.validation == true ? this._qs('#login').disabled = false : this._qs('#login').disabled = true
-                })
-            })
-
-            events.forEach(element => {
-                this._qs("#password").addEventListener(element, () => {
-                    if (this._qs("#password").value == '') {
-                        this._qs('#validation-password').innerHTML = "❌ Enter the password"
-                        this.state.validation = false
-                    } else {
-                        this._qs('#validation-password').innerHTML = ""
-                        this.state.validation = true
-                    }
-
-                    this.state.validation == true ? this._qs('#login').disabled = false : this._qs('#login').disabled = true
-                })
-            })
-
-        }//End of validation
-
-        //Exucute validation
-        validation()
-
-        this._qs('#login').addEventListener('click', () => {
-            this.setLoader();
-            // API call for login
-            const userName = this._qs('#email').value
-            const password = this._qs('#password').value
-            fetch('http://homey-api.atwebpages.com/login', {
-                method: 'POST',
-                headers: {
-                    'Username': userName,
-                    'Password': password
-                }
-            })
-                .then((res) => res.json())
-                .then((res) => {
-                    if (res.data.login === 'true') {
-                        if (this._qs('#remember').checked == true) {
-                            localStorage.login = 'true'
-                            localStorage.token = res.data.token
-                        } else {
-                            sessionStorage.login = 'true'
-                            sessionStorage.token = res.data.token
-                        }
-                        dispatchEvent(new Event('login-success'))
-                        dispatchEvent(new Event('exit-form'))
-                    } else {
-                        localStorage.login = 'false'
-                        localStorage.token = ''
-                        sessionStorage.login = 'false'
-                        sessionStorage.token = ''
-                        dispatchEvent(new CustomEvent("pop-up", { detail: { pop: 'error', msg: res.data.message } }))
-                    }
-                    this.stopLoader();
-                })
-                .catch(err => {
-                    dispatchEvent(new CustomEvent("pop-up", { detail: { pop: 'error', msg: err.message } }))
-                    this.stopLoader();
-                })
-
-        })//End of Login API call
-
-        // Login with Google
-        this._qs('.google').addEventListener('click', async () => {
-            dispatchEvent(new CustomEvent("pop-up", { detail: { pop: 'notice', msg: "Feature disabled at the moment. Use email instead." } }))
-        })
-
-        // Login with FaceBook
-        this._qs('.facebook').addEventListener('click', async () => {
-            dispatchEvent(new CustomEvent("pop-up", { detail: { pop: 'notice', msg: "Feature disabled at the moment. Use email instead." } }))
-        })
 
     }//End of connectedCallback
