+++ src/index.html
@@ -7,7 +7,7 @@
     <title>Homey</title>
     <link rel="icon" href="/favicon.ico" />
     <link rel="stylesheet" href="assets/css/reset.css" />
-    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
+    <script defer src="assets/js/libraries/axios.min.js"></script>
 </head>
 
 <body>
